+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # ChangeLog
 
 ## Version 1.0
+
+- (B.5) Getters/setters for activation functions
 - (B.4) Getters/setters for weights and biases
 - (B.3) Added weights and biases to _feed forward neural network_
 - (B.2) _CELU_ implementation.